boilerplate/pyproject.toml
[tool.poetry]
name = "boilerplate"
version = "0.1.0"
description = ""
authors = ["msm9401 <anstmdals94@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
drf-yasg = "^1.21.7"
django-debug-toolbar = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
